rafs: move RafsSuperFlags into metadata/mod.rs
The RafsSuperFlags are used in metadata/mod.rs, so move it there.
